+++ packages/Peptides/src/utils/constants.ts
@@ -43,6 +43,7 @@ export enum TAGS {
   MULTIPLE_VIEWS = 'isMultipleViews',
   IDENTITY_TEMPLATE = 'Identity template',
   SIMILARITY_TEMPLATE = 'Similarity template',
+  ANALYSIS_COL = 'isAnalysisCol',
 }
 
 export enum SEM_TYPES {