+++ packages/Peptides/src/utils/misc.ts
@@ -36,7 +36,7 @@ export function scaleActivity(activityCol: DG.Column<number>, scaling: C.SCALING
       const val = activityColData[i];
       return val === DG.FLOAT_NULL || val === DG.INT_NULL ? val : formula(val);
     });
-
+  scaledCol.setTag(C.TAGS.ANALYSIS_COL, `${true}`);
   return scaledCol;
 }